refactor(bulk fastt): extract types and helper functions into fseparate ile
